separated the ui testing and api testing
