+++ Group09_Selenium_RestAssured_TestNG_Cucumber/.idea/compiler.xml
@@ -7,6 +7,7 @@
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
         <module name="api-testing" />
+        <module name="Group09_Selenium_RestAssured_TestNG_Cucumber" />
       </profile>
     </annotationProcessing>
   </component>
