+++ Group09_Selenium_RestAssured_TestNG_Cucumber/src/test/java/stepdefinitions/ui/UITestingSteps.java
@@ -1,0 +1,4 @@
+package stepdefinitions.ui;
+
+public class UITestingSteps {
+}
