+++ Group09_Selenium_RestAssured_TestNG_Cucumber/src/test/resources/features/ui/UILogin.feature
@@ -1,0 +1,6 @@
+# Created by pavit at 12/14/2024
+Feature: # Enter feature name here
+  # Enter feature description here
+
+  Scenario: # Enter scenario name here
+    # Enter steps here
